+++ libft/ft_strlcat.c
@@ -14,9 +14,9 @@
 
 size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 {
-	size_t	src_len;
-	size_t	i;
-	size_t	j;
+	unsigned int	src_len;
+	unsigned int	i;
+	unsigned int	j;
 
 	src_len = ft_strlen(src);
 	i = 0;
